Bug 1399141 - Attribute malloc memory when creating reflector object r=baku
UltraBlame original commit: ff90282974cef1e1b73b826f165aa7be26246a6a

+++ dom/file/Blob.h
@@ -151,7 +151,7 @@ private:
 
 
 
-uint64_t BindingJSObjectMallocBytes(Blob* aBlob);
+size_t BindingJSObjectMallocBytes(Blob* aBlob);
 
 } 
 }
